Add PL and fix country for CA
